- Linting - Move ClientConfiguration to root namespace
